[logfile] Fix format string in debug log.

+++ src/logfile.cc
@@ -93,10 +93,10 @@ throw (error)
 
         loo.loo_fd.close_on_exec();
 
-        log_info("Creating logfile: fd=%d; size=%d; mtime=%d; filename=%s",
+        log_info("Creating logfile: fd=%d; size=%" PRId64 "; mtime=%" PRId64 "; filename=%s",
                  (int) loo.loo_fd,
-                 this->lf_stat.st_size,
-                 this->lf_stat.st_mtime,
+                 (long long) this->lf_stat.st_size,
+                 (long long) this->lf_stat.st_mtime,
                  filename.c_str());
 
         this->lf_valid_filename = true;
